Changed way of sending commands to the MC
Instead of mmaping an MC portal to send commands to the MC directly
from user-space, now an ioctl is used, so that we can intercept
in the kernel the commands before they are sent to the MC and
the responses from the MC, before passing them back to user-space.
This is needed to provide a more robust way of doing clean up of MC
state when the resman tool terminates unexpectedly.

Change-Id: I8026ae43338301a1315916840ee99c7effe51a74
Signed-off-by: J. German Rivera <German.Rivera@freescale.com>