remove check of global MC version numbers
Remove the global check that refuses to run restool unless
the MC major version matches.  This is not necessary at
the global level-- restool will be able to do basic dprc
commands with any MC version number.

(What is needed in the future is some additional intelligence
added so that commands for new objects are only executed
if the objects actually exist in the MC firmware)

Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>