added lookup to determine object version for a given MC firmware version
The MC currently does not have a mechanism to advertise which version of an
object the current firmware supports.  All it advertises is the MC firmware
version.  Add a lookup table so that restool can determine which version of
an object it is dealing with and use the appropriate MC command function

Signed-off-by: Gordie White <gordie.white@nxp.com>
[Stuart: fixed typo in commit message]
Signed-off-by: Stuart Yoder <stuart.yoder@nxp.com>