dprc: generate-dpl: export only the objects that exist
When creating the obj_set, the write_obj_set function takes into
consideration only the first_index and the last_index of the object set
without taking into consideration that there are platforms where the
dpmac IDs can be sparse (eg LX2: dpmac.2 dpmac.3 dpmac.17 dpmac.18).

Fix this by splitting the function into multiple ones that could be
called exactly when there are new ids to be added to the current
object set.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>