restool: add global option --rescan
Add the --rescan flag to issue a rescan of the fsl-mc bus
after issuing the MC command.
The default behavior of restool is changed because without this flag,
from now on, an explicit rescan on the fsl-mc bus is not performed.
This is ok because we rely on the MC generated interrupts for object
creation or deletion.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>