ls-main: introduce the get_interface_name function
The newly introduced get_interface_name function would check for the
interface names corresponding to a particular DPAA2 object. With the
latest changes in how we do a rescan of the bus, we could be checking
the sysfs too early (ie before the object probed) thus add a number of
retries before giving up.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>