Does not create DPDBG object in case of help command
The ls-debug helper script automaticaly creates the DPDBG object,
in case it does not exists,  when an argument is given. The problem
is that it also creates the object in case of help commands.
The patch resolve this issue and does not create the DPDBG object
for --help or -h arguments given to ls-debug script.

Signed-off-by: Grigore Popescu <grigore.popescu@nxp.com>