ls-main: fix endpoint parsing
In the circumstance that there are more than 9 endpoints in a dpdmux
object, the parsing code that tries to verify if the endpoint is
already connected fails because it does not take into account the fact
that both 'endpoint 1' and 'endpoint 10' have the same prefix.
Fix this by exactly parsing for the string.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>