scripts: add the ls-delete helper script
We can now know which objects the suppliers for another object. This is
really useful when trying to undo the operations done by a ls-addni
because we can know which objects should be unbind from their respective
driver and destroyed alongside the main one.

Use this functionality in the new ls-delete helper script.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Cosmin-Florin Aluchenesei <aluchenesei.cosmin-florin@nxp.com>
Signed-off-by: Robert-Ionut Alexa <robert-ionut.alexa@nxp.com>