restool: add bash completion script
Add a basic restool bash completion script that lists the restool
commands that can be used on a specific command and also the objects
that are available.

The script is installed in /usr/share/bash-completion/completions/ so
that the default bash-completion package can pick it up directly without
any other setup required.

Signed-off-by: Robert-Ionut Alexa <robert-ionut.alexa@nxp.com>