dpni: restrict the number of queues to maximum 2 * no_cores
The maximum possible number of queues currently supported by
restool is 32, but the MC firmware does not allow more than 16
queues for LS2088, LS1088 and LS2085 platforms. Therefore,
creating a dpni with 32 queues on these boards should not be
possible, since it will return an error in MC.
The ls-addni script no longer supports creating a dpni with more
than 2 * no_cores queues. If this is the case, the default number
of queues (no_cores) will be set.

Signed-off-by: Robert-Ionut Alexa <robert-ionut.alexa@nxp.com>