dpni: don't try to print the MAC address when in promisc mode
In the newest MC version, the DPNI_OPT_NO_MAC_FILTER means that no MAC
table is allocated, thus the DPNI is running in promisc mode.  This
means that there is no MAC primary address and that the command will
fail.  Adjust the restool info command so that it does not issue it in
this case.

The same restriction applies also to the restool dpni update command
which cannot set the primary MAC address of a DPNI when
DPNI_OPT_NO_MAC_FILTER is set.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>