dprc: fix memory illegal access
curr_obj_type array (size 8) is filled with 8 bytes from curr_obj->type which
is an array of size 16. This means that curr_obj_type could be not
null terminated. Further, this array is passed to a strcmp which expects
a null terminated string. Therefore the size of curr_obj_type array is
incremented by one always keeping the last character '\0'.

Signed-off-by: Cosmin-Florin Aluchenesei <aluchenesei.cosmin-florin@nxp.com>