dpdmux: remove unused dpdmux_create_options array
clang-16 reports:

dpdmux_commands.c:95:22: error: variable 'dpdmux_create_options' is not needed and will not be emitted [-Werror,-Wunneeded-internal-declaration]
static struct option dpdmux_create_options[] = {
                     ^
1 error generated.

Remove the dpdmux_create_options array and the enum which now also
becomes unused.

This is a temporary fix which only targets the compile time warning.
Another problem that is left is the fact that the dpdmux_create V10
command uses dpdmux_create_options_v9 instead of having a separate
structure defining all the accepted fields. We cannot just switch over
to using the structure that we are removing in this patch because all
the latest create time options are listed below the _v9 one.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>