dpni: dump policer stats for all Rx TCs
Even though the policer statistics are per Rx traffic class, restool was
ignoring that parameter and printing them only for TC 0.
Extend what statistics are printed by iterating over all the Rx traffic
classes and getting the policer stats. Also, the policer stats are not
printed if the policer is not enabled on a specific TC.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>