dpni: extend congestion stats printing
The congestion statistics that are exported by the MC firmware are,
depending on the configuration of the DPNI, per queue and tc or just per
tc. Even though this infrastructure was available, restool was not
printing all of them.

Extend the congestion stats printing so that all the available stats are
there.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>